TestApi/pyproject.toml
[tool.poetry]
name = "investlink"
version = "0.1.0"
description = ""
authors = ["Сайденов Рамзан <rama050102@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
celery = "^5.3.6"
redis = "^5.0.1"
sseclient = "^0.0.27"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
